requirements.txt
# ScryptBot Requirements
feedparser
python-dotenv
tweepy
openai
yfinance
pandas
playwright
Pillow
